feat: introduce ZTD playground for development loop validation and add `init` and `ztdConfig` CLI commands

+++ playgrounds/ztd-playground/tests/fixture-schema.ts
@@ -22,7 +22,7 @@ export const productSchema: TableSchemaDefinition = {
 export const orderSchema: TableSchemaDefinition = {
   columns: {
     orders_id: 'INTEGER',
-    user_id: 'BIGINT',
+    user_id: 'INTEGER',
     order_date: 'TEXT',
     status: 'TEXT'
   }
